Add space after . and , in ALT description
Since the Teletekst font has the . and the , in the left part of the character cell, a space character is not needed on the TV image (it would actually be nice to have this in a screenshot as well, but here we are...). For a more fluent reading of the ALT description text, the space character is added back, but not when a dot or comma is part of a number.